+++ src/Common.hs
@@ -1,11 +1,13 @@
 module Common
     ( module Common.AbstractError
     , module Common.Env
+    , module Common.Log
     , module Common.Regex
     , module Common.Serialization
     ) where
 
 import Common.AbstractError
 import Common.Env
+import Common.Log
 import Common.Regex
 import Common.Serialization